+++ topic/admin.py
@@ -2,8 +2,9 @@ from django.contrib import admin
 from django.contrib.auth.models import Group, User
 
 # Register your models here.
-from topic.models import Grade, Skill, Topic
-from users.models import Initial_Password, Stage, Your_Stage
+from topic.models import Topic, Skill, Grade
+from users.models import InitialPassword
+from classes.models import Stage, YourStage
 
 admin.site.register(Topic)
 admin.site.register(Skill)
@@ -11,7 +12,7 @@ admin.site.unregister(Group)
 
 
 class StageInline(admin.StackedInline):
-    model = Your_Stage
+    model = YourStage
     can_delete = False
 
 
@@ -34,4 +35,4 @@ admin.site.unregister(User)
 admin.site.register(User, UserAdmin)
 admin.site.register(Grade)
 admin.site.register(Stage)
-admin.site.register(Initial_Password)
+admin.site.register(InitialPassword)
